issue #32 Split Sin class to avoid circular import cause
